V3 (#17)
* Capture History

* attempted capthist fixes (still not working)

* Third Killer Move

* New Net + Datagen Fixes

* Test net + fixed mate scores

* conthist: construction from null is not valid

* Net 004 (it's worse)

* Attempted Conthist Fixes (+ no net 004)

* malus!

* Better malus!

* conthist adaptations  + not clearing!

* Board Optimisations

* Conthist!

* Nets 4 + 5 (+40 +/- 15 elo)

* Futility Pruning!

* Board::makeMove improvements

* Inlining, free 140% nps improvement

* No more compiler error + speedups!

* Actually update nnue lol

* Attempted Goofier Malus

* Working getPV and altered bestMove changes

* Improving (#16)

* Improving

* Working RFP Improving + Untested LMP Improving

* Move ==, And no repitition of killer moves

* Skip Unnecessary SEEs

* no qsearch repitition detection, better history formula

* Proper NNUE Builds

* Better clearHistory & remove unused function

+++ src/tests.cpp
@@ -51,7 +51,7 @@ void runPerftSuite(int number) {
         clock_t end = clock();
         std::cout << "Passed " << std::to_string(passed) << ", Failed " << std::to_string(failed) << ", Skipped " << std::to_string(skipped) << '\n';
         std::cout << "Tests took: " << std::to_string((end-start)/static_cast<double>(1000)) << '\n';
-        std::cout << "Total nodes: " << std::to_string(total) << '\n';
+        std::cout << "Total nodes: " << std::to_string(static_cast<int>(total)) << '\n';
         std::cout << "NPS: " << std::to_string(total / ((end-start)/static_cast<double>(1000))) << '\n';
     }
 }
